requirements.txt
wechatpy[cryptography]
requests-mock
python-alipay-sdk
qcloudsms_py
phonenumbers
